requirements.txt
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
matplotlib==3.8.2
seaborn==0.13.0
flask==3.0.0
flask-cors==4.0.0
ta==0.10.2
scikit-learn==1.3.2
websocket-client==1.6.4
alpaca-trade-api==3.1.1
fredapi==0.5.0
pytz==2023.3
twilio==9.0.0